add categorys metadata

+++ server/src/routes/UpdateData.ts
@@ -5,9 +5,13 @@ import storage from '../units/storage'
 
 
 const UpdateData = async (request: Request, response: Response) => {
+  const { IS_PROD } = process.env
+
+  if (IS_PROD !== 'false')
+    await storage.delete('1C.json')
+  
   await storage.delete('combined.json')
   await storage.delete('contentful.json')
-  await storage.delete('1C.json')
   setTimeout(async () => {
     await useCombinedData()
